move TIME_FORMAT into lsps0
