+++ channel_opener_server.go
@@ -8,10 +8,10 @@ import (
 	"log"
 	"time"
 
-	"github.com/breez/lspd/basetypes"
 	"github.com/breez/lspd/btceclegacy"
 	"github.com/breez/lspd/interceptor"
 	"github.com/breez/lspd/lightning"
+	"github.com/breez/lspd/lsps0"
 	lspdrpc "github.com/breez/lspd/rpc"
 	"github.com/breez/lspd/shared"
 	ecies "github.com/ecies/go/v2"
@@ -146,7 +146,7 @@ func (s *channelOpenerServer) RegisterPayment(
 		pi.OpeningFeeParams = &lspdrpc.OpeningFeeParams{
 			MinMsat:              uint64(node.NodeConfig.ChannelMinimumFeeMsat),
 			Proportional:         uint32(node.NodeConfig.ChannelFeePermyriad * 100),
-			ValidUntil:           time.Now().UTC().Add(time.Duration(time.Hour * 24)).Format(basetypes.TIME_FORMAT),
+			ValidUntil:           time.Now().UTC().Add(time.Duration(time.Hour * 24)).Format(lsps0.TIME_FORMAT),
 			MaxIdleTime:          uint32(node.NodeConfig.MaxInactiveDuration / 600),
 			MaxClientToSelfDelay: uint32(10000),
 		}
