+++ lsps0/time.go
@@ -1,3 +1,3 @@
-package basetypes
+package lsps0
 
 var TIME_FORMAT string = "2006-01-02T15:04:05.999Z"
