+++ shared/opening_service.go
@@ -9,7 +9,7 @@ import (
 	"sort"
 	"time"
 
-	"github.com/breez/lspd/basetypes"
+	"github.com/breez/lspd/lsps0"
 	"github.com/btcsuite/btcd/btcec/v2"
 	"github.com/btcsuite/btcd/btcec/v2/ecdsa"
 )
@@ -47,7 +47,7 @@ func (s *openingService) GetFeeParamsMenu(token string, privateKey *btcec.Privat
 		params := &OpeningFeeParams{
 			MinFeeMsat:           setting.Params.MinFeeMsat,
 			Proportional:         setting.Params.Proportional,
-			ValidUntil:           validUntil.Format(basetypes.TIME_FORMAT),
+			ValidUntil:           validUntil.Format(lsps0.TIME_FORMAT),
 			MinLifetime:          setting.Params.MinLifetime,
 			MaxClientToSelfDelay: setting.Params.MaxClientToSelfDelay,
 		}
@@ -82,9 +82,9 @@ func (s *openingService) ValidateOpeningFeeParams(params *OpeningFeeParams, publ
 		return false
 	}
 
-	t, err := time.Parse(basetypes.TIME_FORMAT, params.ValidUntil)
+	t, err := time.Parse(lsps0.TIME_FORMAT, params.ValidUntil)
 	if err != nil {
-		log.Printf("validateOpeningFeeParams: time.Parse(%v, %v) error: %v", basetypes.TIME_FORMAT, params.ValidUntil, err)
+		log.Printf("validateOpeningFeeParams: time.Parse(%v, %v) error: %v", lsps0.TIME_FORMAT, params.ValidUntil, err)
 		return false
 	}
